services/copilot-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
sqlalchemy[asyncio]==2.0.25
aiomysql==0.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
groq==0.11.0
google-generativeai==0.7.2
openai==1.40.0
pytest==7.4.4
pytest-asyncio==0.23.3
